fix: exposure 걷어내기 전

+++ common_api_server/src/main/java/com/karrotmvp/ourapt/v1/user/repository/UserCustomRepositoryImpl.java
@@ -59,7 +59,7 @@ public class UserCustomRepositoryImpl implements UserCustomRepository<User, Stri
       "SELECT u " +
         "FROM User u LEFT JOIN FETCH u.checkedIn " +
         "ORDER BY u.createdAt DESC", User.class);
-    query.setFirstResult((int) pageable.getOffset());
+    query.setFirstResult(Math.toIntExact(pageable.getOffset()));
     query.setMaxResults(pageable.getPageSize());
     List<User> adminUsers = new ArrayList<>();
     Set<String> normalUserIds = new HashSet<>();
